remove protecting code
